pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "car_image_classification_using_cnn"
version = "0.0.1"
description = "This is the project description for Group 140 in the 02476 Machine Learning Operations course at DTU. The overall goal of the project is to apply the material learned throughout the course to a realistic machine learning problem, with a strong focus on MLOps principles such as reproducibility, experiment tracking, configuration management, and deployment. The project will serve as the basis for the final exam and will be documented through code and a written report."
authors = [
  { name = "klakawaka", email = "your@email.com" },
]
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "fastapi==0.115.6",
  "pillow>=12.1.0",
  "tqdm>=4.67.1",
  "typer==0.15.1",
  "uvicorn==0.34.0",
  "numpy<2",
  # Apple Silicon (M2/M4): modern PyTorch
  "torch==2.6.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
  "torchvision>=0.21.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
  # Intel macOS: last supported wheel line is 2.2.x
  "torch==2.2.2; sys_platform == 'darwin' and platform_machine == 'x86_64'",
  "torchvision==0.17.2; sys_platform == 'darwin' and platform_machine == 'x86_64'",
  # (Optional) Linux/Windows/CI: use modern PyTorch
  "torch==2.6.0; sys_platform != 'darwin'",
  "torchvision>=0.21.0; sys_platform != 'darwin'",
  "pathspec>=0.12.0,<0.13.0",
  "dvc>=3.66.1",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "loguru>=0.7.3",
  "python-dotenv>=1.2.1",
  "wandb>=0.24.0",
  "google-api-python-client>=2.188.0",
  "google-cloud-storage>=3.8.0",
  "dvc-gs>=3.0.2",
  "requests>=2.32.5",
  "python-multipart>=0.0.21",
  "evidently>=0.7.20",
  "scikit-learn>=1.8.0",
  "pandas>=2.3.3",
  "onnx>=1.19.0",
  "onnxruntime>=1.23.2",
  "onnxscript>=0.5.0",
  "bentoml>=1.4.33",
  "ml-dtypes>=0.4.1",
]

[dependency-groups]
dev = [
  "cookiecutter==2.6.0",
  "coverage==7.6.9",
  "invoke==2.2.0",
  "loguru==0.7.3",
  "mkdocs-material==9.7.1",
  "mkdocstrings-python==2.0.1",
  "mypy==1.19.1",
  "pre-commit==4.1.0",
  "pytest==8.3.4",
  "pytest-cov>=6.3.0",
  "ruff==0.1.3",
  "types-tqdm>=4.67.0.20250809",
]

[tool.uv]
# Make the lockfile resolve for macOS and Linux (for Docker)
required-environments = [
  "sys_platform == 'darwin' and platform_machine == 'x86_64'",
  "sys_platform == 'darwin' and platform_machine == 'arm64'",
  "sys_platform == 'linux' and platform_machine == 'x86_64'",
  "sys_platform == 'linux' and platform_machine == 'aarch64'",
]

# Support both macOS and Linux
environments = [
  "sys_platform == 'darwin' and platform_machine == 'x86_64'",
  "sys_platform == 'darwin' and platform_machine == 'arm64'",
  "sys_platform == 'linux' and platform_machine == 'x86_64'",
  "sys_platform == 'linux' and platform_machine == 'aarch64'",
]

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = [
    "tests/*",
    "raw/*",
    "data/*",
    "models/*",
    "outputs/*",
    "reports/*",
    "notebooks/*",
    "*/tmp*/*",
]

[tool.mypy]
ignore_missing_imports = true
